requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
crypto==1.4.1
Django==5.0
django-admin-thumbnails==0.2.8
django-ckeditor==6.7.0
django-js-asset==2.2.0
idna==3.6
Naked==0.1.32
Pillow==10.1.0
pycryptodome==3.19.0
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
